Sets the framework for displaying ranks
